Create tumor_image_classifier_experiment.R
This is a script used to test the image classifier trained in the data_augmentation_image.R script. The model was trained on PanCan cohort data, so has not been added to this git project until I find a replacement public dataset of H&E images to re-train and post.

The outputs include performance statistics and an image plot based on the keras image classifier example that shows correct and incorrect predictions based on image labels (in this case the subfolders they were organized into and pulled from).